requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
drf-spectacular>=0.27,<0.29
